Autoscroll sidebar
Summary:
If the ToC in sidebar is long then the active file can be under the fold when we highlight it.

This also saves some CPU cycles because it highlights only after scrolling of the main window and not in the elements.

Test Plan:
Scrolled on a long diff.
Scrolled the ToC.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, Korvin

Differential Revision: https://secure.phabricator.com/D3473